dom/tests/mochitest/webcomponents/test_document_adoptnode.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1177991
-->
<head>
  <title>Test for Bug 1177991</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1177991">Mozilla Bug 1177991</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

var thrownException = false;
var shadowRoot = document.createElement('a').createShadowRoot();

try {
  document.adoptNode(shadowRoot);
} catch(err) {
  thrownException = err;
}

ok(thrownException !== false, "A HierarchyRequestError");
is(thrownException.name, "HierarchyRequestError", "A HierarchyRequestError should've been thrown");

</script>
</pre>
</body>
</html>
